TelegramRSSFormatter/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "feedparser>=6.0.12",
    "openai>=2.6.0",
    "python-telegram-bot>=22.5",
]
